print various integer size types

+++ 0x00-hello_world/6-size.c
@@ -14,7 +14,7 @@ float f;
 printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
 printf("Size of an int: %i byte(s)\n", (unsigned int)sizeof(i));
 printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(l));
-printf("Size of a long long int: %llu byte(s)\n", (unsigned long long)sizeof(a));
-printf("Size of a float: %lu byte(s)\n", (float)sizeof(f));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(a));
+printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
